refactor(hub): extract dict registry/handle into shape/via-lookup + port seam; retire dead applyLocale; ratchet vault.ts ceiling down (#650)
Task 1 of the via-lookup phase D arc — pure relocation, no behavior change.

- Move DictionaryHandle (+ DictEntry/DictionaryOptions/DICT_COLLECTION_PREFIX/
  dictCollectionName) out of shape/via-i18n/dictionary.ts into
  shape/via-lookup/handle.ts, renamed LookupHandle; dictionary.ts re-exports
  it as DictionaryHandle for compat.
- Extract vault.ts's enforceStaticDictOnPut/resolveDictSource bodies and the
  dictionary() rename closure into pure helpers in
  shape/via-lookup/registry.ts (registry Maps stay vault-resident, passed by
  reference); vault.ts's methods become thin delegators.
- New port/with/lookup-strategy.ts seam (mirrors i18n-strategy.ts) — the
  only lookup module vault.ts imports; kernel spine never reaches
  shape/via-lookup/* directly (Check 14 via-layering).
- shape/via-i18n/active.ts's buildDictionaryHandle now delegates to the new
  shape/via-lookup/active.ts's withLookup().buildLookupHandle — same handle,
  new home.
- Retire vault.applyLocale (dead: zero production callers anywhere in the
  family, superseded by via.present/Collection.applyLocaleToRecord) —
  removed from kernel-api.golden.json with rationale in the task report.
- Re-ratchet vault.ts's kernel-surface ceiling 4088 -> 3941 (the new
  actual); retarget the enclave body-access ratchet entry from
  shape/via-i18n/dictionary.ts to shape/via-lookup/handle.ts (same count, 3).

Verified: focused dict/i18n suites + new lookup-extraction-parity.test.ts
(64/64), full hub suite (464 files / 3721 tests), typecheck, lint,
check:architecture, build, bundle-check (zero leaks, floor delta +0.2%) —
all green. No existing test file edited.
