+++ packages/hub/__tests__/kernel-api.golden.json
@@ -77,7 +77,6 @@
   ],
   "Vault": [
     "abortSchemaCutover",
-    "applyLocale",
     "archive",
     "archivePeriod",
     "assertCanExport",
